refactor: modify web-container thread-pool executor configuration pattern with general mode. (#1015)
