+++ hippo4j-adapter/hippo4j-adapter-web/src/main/java/cn/hippo4j/adapter/web/TomcatWebThreadPoolHandler.java
@@ -18,6 +18,7 @@
 package cn.hippo4j.adapter.web;
 
 import cn.hippo4j.common.constant.ChangeThreadPoolConstants;
+import cn.hippo4j.common.enums.WebContainerEnum;
 import cn.hippo4j.common.model.ThreadPoolBaseInfo;
 import cn.hippo4j.common.model.ThreadPoolParameter;
 import cn.hippo4j.common.model.ThreadPoolParameterInfo;
@@ -163,5 +164,10 @@ public class TomcatWebThreadPoolHandler extends AbstractWebThreadPoolService {
         } catch (Exception ex) {
             log.error("Failed to modify the Tomcat thread pool parameter.", ex);
         }
+    }
+
+    @Override
+    public WebContainerEnum getWebContainerType() {
+        return WebContainerEnum.TOMCAT;
     }
 }
