+++ hippo4j-adapter/hippo4j-adapter-web/src/main/java/cn/hippo4j/adapter/web/UndertowWebThreadPoolHandler.java
@@ -25,6 +25,7 @@ import java.util.Objects;
 import java.util.concurrent.Executor;
 
 import cn.hippo4j.common.constant.ChangeThreadPoolConstants;
+import cn.hippo4j.common.enums.WebContainerEnum;
 import cn.hippo4j.common.model.ThreadPoolBaseInfo;
 import cn.hippo4j.common.model.ThreadPoolParameter;
 import cn.hippo4j.common.model.ThreadPoolParameterInfo;
@@ -154,5 +155,10 @@ public class UndertowWebThreadPoolHandler extends AbstractWebThreadPoolService {
         } catch (Exception ex) {
             log.error("Failed to modify the undertow thread pool parameter.", ex);
         }
+    }
+
+    @Override
+    public WebContainerEnum getWebContainerType() {
+        return WebContainerEnum.UNDERTOW;
     }
 }
