+++ hippo4j-common/src/main/java/cn/hippo4j/common/enums/WebContainerEnum.java
@@ -15,30 +15,10 @@
  * limitations under the License.
  */
 
-package cn.hippo4j.common.spi;
+package cn.hippo4j.common.enums;
 
-import cn.hippo4j.common.executor.support.CustomBlockingQueue;
-
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.LinkedBlockingQueue;
-
-/**
- * SPI whit generic type test.
- */
-public class MyArrayBlockingQueue implements CustomBlockingQueue<Runnable> {
-
-    @Override
-    public Integer getType() {
-        return null;
-    }
-
-    @Override
-    public String getName() {
-        return this.getClass().getSimpleName();
-    }
-
-    @Override
-    public BlockingQueue<Runnable> generateBlockingQueue() {
-        return new LinkedBlockingQueue<>(20);
-    }
+public enum WebContainerEnum {
+    TOMCAT,
+    JETTY,
+    UNDERTOW
 }
